Replace custom tool download method with the one in the tool cache library.
