+++ lib/codeql.js
@@ -12,14 +12,10 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
 Object.defineProperty(exports, "__esModule", { value: true });
 const fs = __importStar(require("fs"));
 const path = __importStar(require("path"));
-const stream = __importStar(require("stream"));
-const globalutil = __importStar(require("util"));
 const toolrunner = __importStar(require("@actions/exec/lib/toolrunner"));
-const http = __importStar(require("@actions/http-client"));
 const fast_deep_equal_1 = __importDefault(require("fast-deep-equal"));
 const query_string_1 = __importDefault(require("query-string"));
 const semver = __importStar(require("semver"));
-const uuid_1 = require("uuid");
 const actions_util_1 = require("./actions-util");
 const api = __importStar(require("./api-client"));
 const defaults = __importStar(require("./defaults.json")); // Referenced from codeql-action-sync-tool!
@@ -142,21 +138,6 @@ async function getCodeQLBundleDownloadURL(apiDetails, variant, logger) {
     }
     return `https://github.com/${CODEQL_DEFAULT_ACTION_REPOSITORY}/releases/download/${CODEQL_BUNDLE_VERSION}/${codeQLBundleName}`;
 }
-// We have to download CodeQL manually because the toolcache doesn't support Accept headers.
-// This can be removed once https://github.com/actions/toolkit/pull/530 is merged and released.
-async function toolcacheDownloadTool(url, headers, tempDir, logger) {
-    const client = new http.HttpClient("CodeQL Action");
-    const dest = path.join(tempDir, uuid_1.v4());
-    const response = await client.get(url, headers);
-    if (response.message.statusCode !== 200) {
-        logger.info(`Failed to download from "${url}". Code(${response.message.statusCode}) Message(${response.message.statusMessage})`);
-        throw new Error(`Unexpected HTTP response: ${response.message.statusCode}`);
-    }
-    const pipeline = globalutil.promisify(stream.pipeline);
-    fs.mkdirSync(path.dirname(dest), { recursive: true });
-    await pipeline(response.message, fs.createWriteStream(dest));
-    return dest;
-}
 async function setupCodeQL(codeqlURL, apiDetails, tempDir, toolCacheDir, variant, logger) {
     try {
         // We use the special value of 'latest' to prioritize the version in the
@@ -205,7 +186,7 @@ async function setupCodeQL(codeqlURL, apiDetails, tempDir, toolCacheDir, variant
                 logger.debug("Downloading CodeQL bundle without token.");
             }
             logger.info(`Downloading CodeQL tools from ${codeqlURL}. This may take a while.`);
-            const codeqlPath = await toolcacheDownloadTool(codeqlURL, headers, tempDir, logger);
+            const codeqlPath = await toolcache.downloadTool(codeqlURL, tempDir, headers);
             logger.debug(`CodeQL bundle download to ${codeqlPath} complete.`);
             const codeqlExtracted = await toolcache.extractTar(codeqlPath, tempDir, logger);
             codeqlFolder = await toolcache.cacheDir(codeqlExtracted, "CodeQL", codeqlURLSemVer, toolCacheDir, logger);
