+++ lib/toolcache.js
@@ -15,6 +15,7 @@ const io = __importStar(require("@actions/io"));
 const actionsToolcache = __importStar(require("@actions/tool-cache"));
 const safeWhich = __importStar(require("@chrisgavin/safe-which"));
 const semver = __importStar(require("semver"));
+const uuid_1 = require("uuid");
 const util_1 = require("./util");
 /*
  * This file acts as an interface to the functionality of the actions toolcache.
@@ -204,6 +205,12 @@ function findAllVersions(toolName, toolCacheDir, logger) {
     }
 }
 exports.findAllVersions = findAllVersions;
+async function downloadTool(url, tempDir, headers) {
+    const dest = path.join(tempDir, uuid_1.v4());
+    const finalHeaders = Object.assign({ "User-Agent": "CodeQL Action" }, headers);
+    return await actionsToolcache.downloadTool(url, dest, undefined, finalHeaders);
+}
+exports.downloadTool = downloadTool;
 function createExtractFolder(tempDir) {
     // create a temp dir
     const dest = path.join(tempDir, "toolcache-temp");
