+++ src/toolcache.ts
@@ -3,10 +3,12 @@ import * as os from "os";
 import * as path from "path";
 
 import * as toolrunner from "@actions/exec/lib/toolrunner";
+import { IHeaders } from "@actions/http-client/interfaces";
 import * as io from "@actions/io";
 import * as actionsToolcache from "@actions/tool-cache";
 import * as safeWhich from "@chrisgavin/safe-which";
 import * as semver from "semver";
+import { v4 as uuidV4 } from "uuid";
 
 import { Logger } from "./logging";
 import { isActions } from "./util";
@@ -222,6 +224,24 @@ export function findAllVersions(
     }
     return versions;
   }
+}
+
+export async function downloadTool(
+  url: string,
+  tempDir: string,
+  headers: IHeaders
+): Promise<string> {
+  const dest = path.join(tempDir, uuidV4());
+  const finalHeaders = Object.assign(
+    { "User-Agent": "CodeQL Action" },
+    headers
+  );
+  return await actionsToolcache.downloadTool(
+    url,
+    dest,
+    undefined,
+    finalHeaders
+  );
 }
 
 function createExtractFolder(tempDir: string): string {
